Omega stop flag, sigma = 1 and minor comments
Introduce a flag to stop the simulation if it's not rotating,
i.e. omega is lower than threshold omegaC. This is check after a minimum
number of steps Nmin. If omega, averaged over print_skip steps, is lower
than threshold for more than x times, stop the MD.
These parameters can appear in the json file, but have defaults.