Internal torque from particles. Cluster and substra modules
Now torque is not finite energy diff but computed as t=r vec F for each
particle and summed. Works and now order of magnitude is more
reasonalbe.

Also switch to shape-sensity etar, sum on r**2, prop to N^2.

Functions manipulating clusters and definine substrates are now in
separate modules, so it's easier to be consistent and scripts are
shorter.