Gaussian pot analysis
I think we might have a problem. With by simply putting Gaussians at the
lattice points, on the border of the unit cells, we have a
discontinuity. Small one, and only on the colloids that really happen to
be there, I'd say a small percentage of the whole cluster. We could
smooth out the tail king of like in the Tanh well shape (test already in
this NB), but that might add weird Fourier components in the potential.